= 9.2.4 = * Some changes in nextend framework
git-svn-id: https://plugins.svn.wordpress.org/nextend-accordion-menu/trunk@857093 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ nextend/cache/cache.php
@@ -12,6 +12,8 @@ class NextendCache{
     
     var $_files;
     
+    var $_text = '';
+    
     var $_cacheTime;
     
     var $_prename;
@@ -46,10 +48,20 @@ class NextendCache{
         $this->_files[] = $file;
     }
     
+    function addText($text){
+        $this->_text.=$text;
+    }
+    
     function getCache(){
-        $time = time();
-        $currentcachetime = $time-$time%$this->_cacheTime;
-        $folder = $this->_path.$this->_prename.$currentcachetime.DIRECTORY_SEPARATOR;
+        if(count($this->_files) == 0 && $this->_text == '') return false;
+        if($this->_cacheTime == 'static' || $this->_cacheTime == 0){
+            $folder = $this->_path.'static'.DIRECTORY_SEPARATOR;
+            $currentcachetime = 0;
+        }else{
+            $time = time();
+            $currentcachetime = $time-$time%$this->_cacheTime;
+            $folder = $this->_path.$this->_prename.$currentcachetime.DIRECTORY_SEPARATOR;
+        }
         $this->createCacheSubFolder($folder, $currentcachetime);
         $hash = $this->createHash();
         $cachefile = $folder.$hash.'.'.$this->_filetype;
@@ -59,6 +71,7 @@ class NextendCache{
             for($i = 0; $i < count($this->_files); $i++){
                 $cached.= $this->parseFile(NextendFilesystem::readFile($this->_files[$i]),$this->_files[$i], $i);
             }
+            $cached.= $this->_text;
             NextendFilesystem::createFile($cachefile, $this->parseCached($cached));
             if($this->_gzip){
                 $php = '<?php '
@@ -66,7 +79,7 @@ class NextendCache{
                         . 'if (extension_loaded("zlib") && (ini_get("output_handler") != "ob_gzhandler")) {'
                         . 'ini_set("zlib.output_compression", 1);'
                         . '}'
-                        . 'readfile("'.$cachefile.'");';
+                        . 'readfile("'.str_replace('\\','/',$cachefile).'");';
                 NextendFilesystem::createFile($cachefilegzip, $php);
             }
         }
@@ -86,11 +99,13 @@ class NextendCache{
     
     function createCacheSubFolder($path, $currentcachetime){
         if(NextendFilesystem::existsFolder($path)) return;
-        $previouscachetime = $currentcachetime-$this->_cacheTime;
-        $remove = NextendFilesystem::folders($this->_path);
-        if($remove !== false){
-            for($i = 0; $i < count($remove) && $remove[$i] != $this->_prename.$previouscachetime; $i++){
-                NextendFilesystem::deleteFolder($this->_path.$remove[$i]);
+        if($this->_cacheTime != 'static' && $this->_cacheTime != 0){
+            $previouscachetime = $currentcachetime-$this->_cacheTime;
+            $remove = NextendFilesystem::folders($this->_path);
+            if($remove !== false){
+                for($i = 0; $i < count($remove) && $remove[$i] != $this->_prename.$previouscachetime; $i++){
+                    if($remove[$i] != 'static') NextendFilesystem::deleteFolder($this->_path.$remove[$i]);
+                }
             }
         }
         if(NextendFilesystem::createFolder($path)) return;
@@ -103,7 +118,7 @@ class NextendCache{
         for($i = 0; $i < count($this->_files); $i++){
             $hash.=$this->_files[$i].filemtime($this->_files[$i]);
         }
-        return md5($this->parseHash($hash));
+        return md5($this->parseHash($hash).$this->_text);
     }
     
     function parseHash($hash){
